pyproject.toml
[tool.poetry]
name = "wiggy-conversion"
version = "0.1.0"
description = ""
authors = ["Michael Brown <kurosama112@gmail.com>"]
readme = "README.md"
packages = [{include = "wiggy_conversion"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pandas = "^2.0.1"
matplotlib = "^3.7.1"
seaborn = "^0.12.2"
pydantic = "^1.10.8"
redis = "^4.5.5"
sqlalchemy = "^2.0.15"
redis-orm = "^0.4.1"
metatrader5 = "^5.0.45"
redis-om = "^0.1.2"
sqlalchemy-mixins = "^2.0.0"
scipy = "^1.10.1"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"
py-linq = "^1.4.0"
fastapi-sqlalchemy = "^0.2.1"
pyjwt = "^2.7.0"
jwt = "^1.3.1"
itsdangerous = "^2.1.2"
faker = "^18.9.0"
passlib = "^1.7.4"
fastapi-camelcase = "^1.0.5"
psycopg2-binary = "^2.9.6"
sentry-sdk = {extras = ["fastapi"], version = "^1.24.0"}
tqdm = "^4.65.0"
bs4 = "^0.0.1"
beautifulsoup4 = "^4.12.2"
mailgun = "^0.1.1"
urllib3 = "^2.0.2"
simplejson = "^3.19.1"
requests = "^2.31.0"
mplfinance = "^0.12.9b7"
yfinance = "^0.2.18"
ta = "^0.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
